Move browser files to WORKDIR only after all fallible downloads
Fixes bug where downloading a new stable could replace the artifact in
$WORKDIR/, and then a download of the alpha could fail,
aborting the script.
but since the WORKDIR cache of stable was already replaced, subsequent
runs would not trigger a notification to the repository builder, as
the stable artifact in WORKDIR would be identical at this point.

Saving the move to WORKDIR until last prevents this hickup