Implement new debug commands: `relay disable` and `relay enable`
Add two new `mullvad debug` subcommands:

- mullvad debug relay enable <country|city|hostname|openvpn|wireguard>
- mullvad debug relay disable <country|city|hostname|openvpn|wireguard>

These commands are used to update the state of relays the current relay
list. This is useful to mock relays going offline or coming online from
an offline state.

These new debug commands were conceived during the development of the
feature for adding warnings for the upcoming OpenVPN deprecation, as
there wasn't a convenient way of mocking this.