Fix nav drawer wrong focus on Android TV
When switching locations on Android TV when the app goes back to the
connect screen, focus was sometimes incorrectly put on the navigation
drawer, which caused it to expand when it shouldn't.

This commit fixes that by detecting when we navigate back to the connect
screen from the select location screen, and when that is the case we
manually focus the select location button.

We check if we get a FocusDirection.Down event to detect if we navigated
back from the select location screen, which isn't optimal, but I
couldn't find a better way to detect this. This works because if we
return from the account or settings screens we get a
FocusDirection.Enter event.