Add CI job to check permissions
This added check is intended to catch unwanted
permissions changes introduced by dependencies.

It was added to the reproducible/F-Droid workflow
since that's the only place where we build release
artifacts at the time of adding this job. However,
it should be moved to a more suitable place once
we expand and refactor our CI setup.