Add nix devshell actions build
This commit adds a job that runs a nix devshell build
as part of the reproducible workflow, however at this
point it will NOT check for reproducibility but rather
just print the checksum of the build. This is because
there are some remaining reproducibility issues that
need to be addressed.