Fix ktor parsing error
App module had transitive dependencies using a older version at runtime
of serializationx. However, when compiling the test:e2e module a later
version than 1.8.0 was used through ktor. This caused deserialization
in runtime to go looking for a method that didn't exist. By putting an
explicit version in the app module we make sure to use a version that
is compatible with ktor.