Convert relay IP tests to use main ROM and expect polling methods
When we extract the in-IP as a string the expect will only evaluate
the condition once, and then fail. However the in-IP might not have
updated at the time the condition is checked, and as such the expect
will fail.

This change makes it possible to re-evaluate using Playwright's
expect polling functionality.