Update v12 settings migration to conditionally insert entry filters
As the v11 settings migration was broken on Android, the v12 settings
migration might have never run before. With the patch to the v11
settings migration in place the v12 settings migration will then run,
which is fine in itself, but the assumptions that were made when
writing the v12 settings migration are now not valid anymore.

The v12 settings migration now has to take into account that the
keys for the filters (`entry_providers` and `entry_ownership`) might
already exist. As such the v12 settings migration has been updated to
only insert the aformentioned keys if they do not already exists in
order to not accidentally wipe a user's existing setting.