Force initial canvas width to be 320px
When rendering on Linux under wayland there is currently a bug in
Electron where the window's width is reported incorrectly for the first
second and is reported as 232px instead of 320px.

Relying on the window's width causes the Map to get rendered too small
for the first rendering and causes it to get stuck there, as after the
first render we don't update the canvas' size, and as such it can't
be rectified when the resize event is fired. This means we have to set
the correct canvas size on the first render, and as such we use the
static value of 320px.