Add blue background color to prevent white flash before first render
This background color is added because when the app is started it is
created with a white background until the React app has mounted and
rendered. This causes a white flash which appears for < 100
milliseconds and is somewhat unpleasant. By setting the background
color to the same blue used for the background in the app we can
ease the transition a bit better.

This white flash was especially noticable on Linux when running under
Wayland.